Backend Finalizado

+++ backend/controllers/tareasController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-Handler') 
-  const Tarea = require('../model/tareasModel')
+const Tarea = require('../model/tareasModel')
+
 
 const  getTareas =asyncHandler( async (req, res) => {
     //const tareas = await Tarea.find({color: "azul"}) --> con un valor en específicio
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
    
     res.status(200).json(tareas)
 
@@ -16,9 +17,9 @@ const  setTareas =asyncHandler( async (req, res) => {
         throw new Error("Por favor teclea una descripción")
         
     }
-
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto, 
+        user: req.user.id
     })
 
     res.status(201).json({tarea})
@@ -31,8 +32,16 @@ const  updateTareas =asyncHandler( async (req, res) => {
         res.status(400)
         throw new Error("La tarea no fue encontrada")    
     }
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
-    res.status(200).json(tareaUpdated)
+
+    //Verificar que la tarea pertenezca al suuario del token
+    if(tarea.user.toString()!== req.user.id){
+        res.status(401)
+        throw new Error("Acceso no autorizado")
+    }else{
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+        res.status(200).json(tareaUpdated)
+    }
+    
 })
 
 const  deleteTareas =asyncHandler( async (req, res) => {
@@ -43,9 +52,16 @@ const  deleteTareas =asyncHandler( async (req, res) => {
         throw new Error("La tarea no fue encontrada")    
     }
 
-    await Tarea.deleteOne(tarea)
-    //await Tarea.findByIdAndDelete(req.params.id)
-    res.status(200).json({id: req.params.id})
+    //Verificar que la tarea pertenezca al suuario del token
+    if(tarea.user.toString()!== req.user.id){
+        res.status(401)
+        throw new Error("Acceso no autorizado")
+    }else{
+        await Tarea.deleteOne(tarea)
+        //await Tarea.findByIdAndDelete(req.params.id)
+        res.status(200).json({id: req.params.id})
+    }
+    
 })
 
 module.exports ={
